refactor(baseball): 상수 리팩토링, 메소드 추출, 애매한 이름 변경
