+++ src/main/java/baseball/Application.java
@@ -1,10 +1,11 @@
 package baseball;
 
+import baseball.constant.BaseballConst;
+import baseball.controller.BaseballGameController;
 import baseball.domain.BaseballGame;
 
 public class Application {
 	public static void main(String[] args) {
-		// TODO: 숫자 야구 게임 구현
 		BaseballGame baseballGame = new BaseballGame();
 		BaseballGameController gameController = new BaseballGameController(baseballGame);
 
@@ -12,6 +13,6 @@ public class Application {
 		do {
 			gameController.gameStart();
 			select = gameController.selectAfterEnd();
-		} while (select == 1);
+		} while (select == BaseballConst.RETRY);
 	}
 }
