+++ src/main/java/baseball/constant/BaseballConst.java
@@ -2,10 +2,14 @@ package baseball.constant;
 
 public class BaseballConst {
 	public static final int NUM_LEN_LIMIT = 3;
-	public static final String INPUT_INTRO = "숫자를 입력해주세요 : ";
+	public static final int RETRY = 1;
+	public static final int EXIT = 2;
+
 	public static final String BALL = "볼";
 	public static final String STRIKE = "스트라이크";
 	public static final String NOTHING = "낫싱";
+
+	public static final String INPUT_INTRO = "숫자를 입력해주세요 : ";
 	public static final String GAME_END = "3개의 숫자를 모두 맞추셨습니다! 게임 종료";
 	public static final String RETRY_OR_EXIT = "게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.";
 }
